requirements.txt
langchain-openai
langchain-community
pymilvus
pymysql
python-dotenv
zhipuai
